Moving scarecrow effect range calculation to tools class
